[frontend] Extract BaseVcsRootComboBox & avoid .*Impl deps
