+++ frontendUiApi/build.gradle
@@ -1,3 +1,7 @@
 vavr()
 
+dependencies {
+  api project(':branchLayoutApi')
+}
+
 addIntellijToCompileClasspath(withGit4Idea: true)
