+++ frontendUiApi/src/main/java/com/virtuslab/gitmachete/frontend/ui/api/vcsrootcombobox/IVcsRootComboBoxFactory.java
@@ -1,0 +1,7 @@
+package com.virtuslab.gitmachete.frontend.ui.api.vcsrootcombobox;
+
+import com.intellij.openapi.project.Project;
+
+public interface IVcsRootComboBoxFactory {
+  BaseVcsRootComboBox create(Project project);
+}
