+++ frontendUiVcsRootComboBoxImpl/build.gradle
@@ -1,0 +1,10 @@
+vavr()
+
+dependencies {
+  implementation project(':logging')
+
+  api project(':frontendUiApi')
+}
+
+addIntellijToCompileClasspath(withGit4Idea: true)
+applySubtypingChecker()
