+++ settings.gradle
@@ -19,6 +19,8 @@ include 'frontendGraphApi'
 include 'frontendGraphImpl'
 include 'frontendIcons'
 include 'frontendUiApi'
+include 'frontendProviderServices'
 include 'frontendUiTableImpl'
+include 'frontendUiVcsRootComboBoxImpl'
 include 'frontendUiRootImpl'
 include 'uiTests'
